finalized

+++ generic-template/runtime/src/types.rs
@@ -40,7 +40,7 @@ pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::Account
 pub type Balance = u128;
 
 /// Identifier of an asset
-pub type AssetId = u128;
+pub type AssetId = u32;
 
 /// Index of a transaction in the chain.
 pub type Nonce = u32;
